Patch to return un-HTMLEscap()'d code from Javascript tags for which the type field reads "text/..." where '...' is something-or-other commentary.  This allows programs active on the client to access the contents of a script tag.  This facility is necessary because it's become a convention to use <script type="text/template"> to store templates for use with client-side libraries like underscore, mustache, and the Microsoft jQuery Template Engine.
I can't take credit for this: tmpvar (the original owner of jsdom)
gave me the initial fix, although I did fix the text value test.

Using "text/..." as the test for whether or not JSDom should return
the contents of the script tag unchange feels like a hack.  But it's
a hack that works in convention, so I think we should live with it
for now.