StylePropertyReflectsStyleAttribute test: Expect non-existent .style property to be undefined instead of ''.
As per browser behavior we should aim for non-existent properties to have a value of '' at some point when someone adds getters and setters for each property to CSSOM. In this particular test, however, the 'color' property was only '' because it had been defined previously, ie. an artifact of the old implementation.