Update Sizzle tests to the latest from GitHub.
This reveals a number of bugs in jsdom and nwmatcher. The latter are filed.

I tried to minimize the modifications to the Sizzle test suite using some new techniques in the runner. The only remaining ones are needing to change the `broken` function declaration to a variable expression, and commenting out all the focus-related stuff since we don't have `focus` and `blur` on all elements (yet).