Consolidate all jQuery fixtures used by the tests.
It was kind of silly how scattered and different they all were. This preserves all the different versions, but puts them in one place. And I downloaded the current version at code.jquery.com, version 1.11.0, and used that where we were currently unnecessarily going to the internet.

While I was there, rewrote the JSONP test to be in the new style, since it was small and easy.