Add test to show inconsistency with jsdom.env
jsdom.env is normally used to be the easiest access point to parse a
website. However it doesn't load external scripts by default.
By supplying jsdom.env with the proper features config it is possible to
load external scripts, however, some bug in jsdom.env prevents that the
referenced external script is loaded an processed. If, however, a src or
scripts reference is supplied with the config, the referenced external
scripts on that page are also executed.