Improve the performance of define by about ~40%
Since define is used to copy over all of our dom properties (which are a lot), it has to be pretty performant.
This patch does two things:
* Removes the need to allocate a new anonymous function by replacing the forEach with a standard for loop
* If the property to copy is a simple defined property (not defined with defineProperty), avoid the call to defineProperty ourselves too

Especially the second point is important. It shaves off about 40% of time in define, and about 20% of overall initialization time, which reduces the time for 10000 instances from about 8 seconds to 6 seconds on my laptop.