replaced grave accents with common single quotes
The grave accent '`' (quoting a accent is weird...) was causing nodejs to crash .
```
jsdom.js:55
    throw new RangeError(`Invalid parsingMode option ${JSON.stringify(options.
                         ^
SyntaxError: Unexpected token ILLEGAL
[...]
```
Replaced it with a "normal" quote and escaped the already existing quotes. (Imho it would be better to use single quotes for strings and double quotes for actual quotes in strings, but since all strings are declared using double quotes, I adapted the code style).

There is still another issue with node concerning sets, but this seemed easier to fix.