Introducing NamedPropertiesTracker
This is a helper class used to mimic the default getters in the html specification that are used on objects which receive named properties (Window, Document and <form>).
It is not a complete solution, because that would require something like ES6 Proxy. To be specific, the required getter for a named property is not defined on an object property which was *just* removed using `delete`.