Script errors are no longer part of initialization
As such, we change the `load(errors, window)` callback to `onload(window)`, and now `done` only takes a single error (like a normal program!), instead of an array of errors. But, we do proxy any script *loading* errors (as opposed to execution errors) to the virtual console.

Maybe we should do that with all resources, actually? In resourceLoader, for example?