Always run web platform tests from a server
This fixes certain base URL tests I am soon hoping to add, which use hrefs like "/foo/bar.html" that don't work very well on the file system.

In theory we should be running their python server. But, it requires modifying your hosts file, which is pretty sad. We can just do this hack until we run in to the need to run tests that require that level of complexity.