Properly implement document#createEvent
Due to a bit of an unclarity in the spec I assumed an empty string event type marks an uninitialized event, which is not actually the case. Instead, an event is always initialized if instantiated directly.

The only way a page can create an uninitialized script is via document.createEvent, which now marks events uninitialized via privatedata explicitly.