Correct the logic as to uppercase if NOT xhtml AND html.
Note that currently this breaks many, many tests.

From MDN:

In XML (and XML-based languages such as XHTML), tagName preserves case.
On HTML elements in DOM trees flagged as HTML documents, tagName returns
the element name in the uppercase form. The value of tagName is the same
as that of nodeName.