Named access nested context revision
This updates the "named access on the window object" feature to only
return the nested browsing context (the `contentWindow`) when the
browsing context name is set. Since the browsing context name is not set
by the `id` property, the iframe element should be returned when the id
is the name being used.

In summary:

* `id`: return the element.
* `name`: return the WindowProxy.

This is based on findings from a Firefox issue:
https://bugzilla.mozilla.org/show_bug.cgi?id=1281604