Overhaul HTMLCollection usage
The previous commit updated HTMLCollection to work on impls as input, to avoid a lot of confusing back and forth, but this necessitates a lot of code updates. Along the way, this gets rid of the inefficient/general-purpose mapper() utility in favor of more direct and specialized domSymbolTree manipulation.

In the course of cleaning this all up, various collection-returning getters (such as `document.applets` or `table.cells`) were made slightly more spec-compliant and efficient. Also added `document.scripts`, `document.embeds`, and `document.plugins`.