Further fixes in light of HTMLCollection changes
Notably, we were not correctly updating the version of a node if its children or ancestors changed! This caused our invalidation logic to be very off, unless we used the document as the versioning root (which, previously, we often did).

Some of the older tests were wrong; replaced with to-upstream versions with corrected semantics.