Move lastModified property from Blob class to File
It in fact belongs to File in the spec. See [1] vs [2].

Also changed was the semantics of the property, in case lastModified was
not provided as part of the options dictionary. The specification
dictates that if not provided, the lastModified property must be equal
to Date.now() at the time when the constructor is called (see [3]).

[1]: https://w3c.github.io/FileAPI/#blob
[2]: https://w3c.github.io/FileAPI/#file
[3]: https://w3c.github.io/FileAPI/#file-constructor