Port File API to Web IDL
This uses the new-in-webidl2js [WebIDL2JSFactory] technology to port FileReader, Blob, and File to the webidl2js system. Along the way, it fixes a lot of other correctness issues, in addition to the correctness gains that are automatic from the Web IDL bindings. Notably, the size property on Blob returns 0 when the Blob is closed, and slice()'s algorithm now works much better for unusual-but-valid inputs (such as negative numbers, or non-numbers).

See #1650 for an earlier draft of this.