Delete webkit prefixes and other internal selectors in default CSS
See tmpvar/jsdom#1560.

In particular, this helps because it takes out webkit prefixes in selectors,
which leads to much fewer exceptions being thrown.

I've tested this patch by running the test that runs slowly with breakpoints
set on "all exceptions". With this change, the only exceptions nwmatcher throws
are related to case insensitive attribute matchers (fixed in the next commit)
and "@namespace "http://www.w3.org/1999/xhtml";

This commit and the next commit improve performance in one of my tests from 28s
to 8s.