Start web platform tests server before tests run
This is an attempt to address recent flaky builds, which we think might be caused because the server is starting in the middle of a JSDOM API test, causing a timeout for the currently-running JSDOM API test.

The previous design started booting the server the first time it was needed, and made all tests wait for the server-creation promise before commencing their actual logic. Now, server-creation runs in a before() block.

In the process, it removes the fallback to w3c-test.org, as that configuration isn't tested and generally can cause issues.