Use `this.setTimeout` in window.postMessage
Following in the same vein as #1312 (discussed there and at https://github.com/facebook/jest/issues/625), this makes the same change for the implementation of `window.postMessage`. 

As it was, the same Jest issue applied here when trying to mock timers before posting messages.  Adjusting the reference to `setTimeout` solves the issue, allowing Jest's timer mocking to succeed.