Introduce a new resource loading API
This is a first draft at the new resource loading API discussed in #2050. For now the old resource loading API, and the old API in general, are preserved; they will be removed in subsequent commits.

This also raises the minimum required version to Node.js v8, as the new resource loader will be part of a major release anyway, and the maintenance burden of supporting that old vm module is getting too high.