Drop unused code.
This function is never called with a `node.type` set to `"script"`, `"style"` or
`"root"`. Having those unused cases hang around is just confusing. Moreover, the
code meant do special processing for template elements does not appear to be
correct. When someone decides to tackle it, it should be rethought from the
ground up rather than rely on old untested code.