Move from sax to saxes.
This is a breaking change in the following sense. JSDOM used to rely on sax, but
sax did not report all malformedness errors. One prominent case is the empty
string. An empty string is not a well-formed XML document, but sax accepted
it. This commit replaces sax with saxes. Saxes implements many more checks than
sax, and will thus report errors that sax does not. The upshot of this change is
that JSDOM will report errors on XML documents it previously accepted.

Closes #2213