canvas/canvas-prebuilt: show error when module is available but fails to load
In some environments, neither `canvas`, nor `canvas-prebuilt` will be installed - as it is an optional dependency, we will just ignore their absence, however when one or both of them are installed, there is a chance that loading them fails, as they are depending on native bindings.

For example when building a docker image that runs `sid`, `libpng12-0` is not available, which means that loading (requiring) canvas-prebuilt fails, due to:

> error while loading shared libraries: libpng12.so.0: cannot open shared object file: No such file or directory

It took me quite a while to figure out why canvas-prebuilt worked outside the docker container, but not inside, even though both systems were linux.
The way the try/catch is currently set up means that loading errors are swallowed. The change in this PR makes loading changes at least visible in a similar manner to how they would show up when just doing `require('canvas-prebuilt')` anywhere in the bootstrap code.