Fix XHR sending empty request body after a preflight request
A race condition in the implementation of XMLHttpRequest causes the request
body to be set to undefined after the preflight request, but before the actual
request is sent.

This commit simply removes the cleanup code causing the issue. As
XMLHttpRequest represents single requests, manual cleanup should not be
necessary.

Fixes #2300.