Let canvas/canvas-prebuilt require() errors bubble out
Even if canvas or canvas-prebuilt are installed, there is a chance that loading them fails, as they are depending on native bindings.

For example when building a docker image that runs sid, libpng12-0 is not available, which means that requiring canvas-prebuilt fails, due to:

> error while loading shared libraries: libpng12.so.0: cannot open shared object file: No such file or directory

This change makes it so that such errors surface to the user, instead of getting swallowed (i.e. treating errors loading them in the same way as them not existing).