Fix getting value of Select when selectedIndex is -1
When a doing get value of Select element with selectedIndex=-1, it
should return empty string "" instead of value of its first option.

We update the implementation of `get value() {}` to follow the wording
of whatwg spec.

Also, when Select's `selectedIndex` was modified, we update the Option
elements' `_selectedness` and `_dirtyness` instead of theirs `selected`
attribute. Changing an Option's `selected` atrribute causes parent
Select element to be reset, which makes Select's `selectedIndex` to be
changed to a non `-1` value.

References:
https://html.spec.whatwg.org/multipage/form-elements.html#dom-select-selectedindex
https://html.spec.whatwg.org/multipage/form-elements.html#dom-select-value