Update w3c-xmlserializer dependency
This updates to the latest version of w3c-xmlserializer, which no longer contains a generated XMLSerializer class, but just a serialize function. Instead, we now wrap that package's serialize function as appropriate. This allows us to properly generate an XMLSerializer constructor per jsdom Window, in line with other recent constructor and prototype reform.