Make test/api/encoding use fs.promises
The `pn` package was removed as a dependency in
d240291edbe7d4180a5152993ced7950c834dc57, but it turns out that we are
still using it in test/api/encoding.js. Remove this very last usage of
it. (CI still works as it was included through webidl2js, but webidl2js
has also removed pn as a dependency recently.)

Fixes #2910.

Co-authored-by: Timothy Gu <timothygu99@gmail.com>