Better activation behavior for <label> elements
This commit deals with two cases where the current activation behavior
for <label> elements is deficient:

- When the target of the event that triggered said activation behavior
  is in fact part of some interactive content descendant of the <label>,
  then nothing should be done.

- When the target of the event is an inclusive descendant of the labeled
  control, then the control presumably has already received this event,
  so nothing more should be done by the <label> element.

The former case is specifically treated by the spec as a no-op case,
under https://html.spec.whatwg.org/C/#the-label-element. This applies to
such trees as

    <label>
      <button>Click me!</button>
    </label>

Since the <button> is labelable, if upon clicking the button the <label>
element tries to click the button once again, then an infinite recursion
would result. This is indeed the essence of #2717.

The latter case is not mentioned in the spec directly, but is
implemented by all major browsers. Consider instead the following tree:

    <label>
      <progress>Click me too!</progress>
    </label>

The <progress> element is not interactive, and therefore does not fall
under the jurisdiction of the previous case. But it is still labelable,
and is the labeled control of the <label> element. So when the
<progress> element is clicked, the <label> should also avoid trying to
click the <progress>; otherwise, infinite recursion would emerge here as
well.

It may first appear that the first case is a special case of the second;
indeed, the example above with a single button would fall into the
second case also. But this is not true in general. Consider instead:

    <label>
      <button>Don't click me!</button>
      <button>Click me!</button>
    </label>

If the second button is clicked, then this falls into the first case but
not the second. But we still need to prevent the <label> from clicking
the first button (the actual labeled control) instead.

Fixes #2717.