Remove the dependency on request
Closes #2792. This reimplements the relevant functionality in the agent-factory.js and http-request.js helper files.

The new HTTP stack implementation fixes several previously-failing XMLHttpRequest tests, where the request module was previously not flexible enough to let us manipulate the headers appropriately.

Note that request remains as an indirect dev dependency of the wd package. This also introduces a dependency on the form-data package which we would like to remove in the future.