Be robust against global polyfills
If we are running on a JavaScript engine that does not natively support
a global in lib/jsdom/browser/js-globals.json, but the user loaded a
polyfill prior to loading jsdom, then we could be tricked into assuming
that global is available in new VM contexts.

Do not assume such a thing, and always check for each global's
availability in the VM context.

One could test this by running the following script on Node.js v14.x,
which does not implement AggregateError natively:

    global.AggregateError = class AggregateError extends Error {};
    const { JSDOM } = require("jsdom");
    new JSDOM("", { runScripts: "dangerously" });

Fixes #3205.