Make JSDOM.reconfigure() adjust the History API with the updated URL
Previously, after using JSDOM.reconfigure() to change the URL, the
History API held onto a cached version of the previous URL. This means
if one used the History API like:

    window.history.replaceState(null, "");

The cached URL would be used to replace the value of
window.location.href back to the original. This is unexpected. Not
supplying a third argument should not change the URL.

To fix, update the current history entry when calling reconfigure().

Fixes #3504