MouseEvent: fix {page,offset}{X,Y} during dispatch
Ensure that pageX and pageY reflect the coordinate relative to the origin of the document, i.e. clientX/clientY in the absence of scrolling support, during dispatch.

Also fix the target-relative coordinates, offsetX and offsetY, during dispatch. These will typically be (0, 0), e.g. in a synthetic click or a dispatched MouseEvent with default coordinates, but it will mirror clientX/clientY if those are set.

Lastly, remove some scrollX/scrollY logic for values computed outside of dispatch. While it logically followed the spec, scrolling is not implemented and it is unsafe to base the calculation on those replaceable properties.