Update documentation to say canvas 3.x, not 2.x
As of this commit, jsdom's package.json refers to canvas 3.x. I've modified this part of the readme to mention 3.x. I also removed mention of 2.x and 1.x, but I don't know if that is a desirable change. I wouldn't know what to replace it with anyway.

If a different modification is preferred, consider this PR as a reminder to change the docs; feel free to rewrite my PR.