Cache the document's base URL and serialized base URL
This accounts for a chunk of time in running large tests, because asserts always look at HTMLScriptElement's src.

As part of this we move all the logic out of a helper file and into methods on the Document impl class.

This also discovers and documents (but does not fix) a bug introduced by our current base URL computation process.

Part of #3154.