Split "npm test" into more subcommands
Helps with #3154 by creating a new Node process for WPTs, separate from all the other tests.

Get rid of the index.js file that was necessary to list all the tests.

Move test/jquery-fixtures into test/to-port-to-wpts/ since those are the only tests that use it.

Update npm scripts to follow the convention of prefix:subcommand.