Update dependencies and dev dependencies
This notably includes a cssstyle upgrade which comes with several correctness fixes.

In the course of this upgrade, it became clear several of our existing to-upstream tests do not match browsers.

* Update getComputedStyle-mutations.html to use px instead of em, since we don't correctly resolve em to px
* Update stylesheet-appropriate-time-to-obtain.html to properly catch errors, and to test for the default color for body instead of the empty string. This required new failure expectations.
* Update no-browsing-context.html to test for the default color for body instead of the empty string. This required new failure expectations.

See #3887 for more investigation into broken to-upstream tests, and https://github.com/jsdom/cssstyle/issues/212 for one cssstyle regression.